+++ Sources/KeyboardKit/_Deprecated/KeyboardLocale+Deprecated.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public extension KeyboardLocale {
+    
+    @available(*, deprecated, renamed: "sami_inari")
+    static var inariSami: Self { .sami_inari }
+    
+    @available(*, deprecated, renamed: "sami_northern")
+    static var northernSami: Self { .sami_northern }
+}